# phase-2-fullstack-todo-app/specs/001-task-management-api/contracts/api-spec.openapi.yaml
openapi: 3.0.3
info:
  title: Task Management Backend API
  description: Secure REST API for managing user-specific todo tasks
  version: 1.0.0
  contact:
    name: Task Management API Team
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.taskmanager.example.com/api
    description: Production server
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with credentials and return JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current session and clear authentication token
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User's email address
        name:
          type: string
          maxLength: 100
          description: Optional display name
        created_at:
          type: string
          format: date-time
          description: User creation timestamp
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          minLength: 8
          description: User's password
    LoginResponse:
      type: object
      required:
        - user
        - token
        - expires_at
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT access token
        expires_at:
          type: string
          format: date-time
          description: Token expiration timestamp
    LogoutResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: Logout success indicator
        message:
          type: string
          description: Success message
...
    ErrorResponse:
      type: object
      required:
        - error
        - status_code
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              enum:
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_ERROR
                - INTERNAL_ERROR
                - INVALID_TOKEN
                - TASK_OWNERSHIP_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties:
                type: string
              description: Field-specific validation errors
        status_code:
          type: integer
          description: HTTP status code
          enum: [400, 401, 403, 404, 500]